Add CSRF protection for the customer

+++ src/main/java/com/zuehlke/securesoftwaredevelopment/controller/CustomerController.java
@@ -2,6 +2,7 @@ package com.zuehlke.securesoftwaredevelopment.controller;
 
 import com.zuehlke.securesoftwaredevelopment.config.AuditLogger;
 
+import com.zuehlke.securesoftwaredevelopment.config.CsrfHttpSessionListener;
 import com.zuehlke.securesoftwaredevelopment.domain.Address;
 import com.zuehlke.securesoftwaredevelopment.domain.CustomerUpdate;
 import com.zuehlke.securesoftwaredevelopment.domain.NewAddress;
@@ -9,9 +10,12 @@ import com.zuehlke.securesoftwaredevelopment.domain.RestaurantUpdate;
 import com.zuehlke.securesoftwaredevelopment.repository.CustomerRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpSession;
 
 @Controller
 
@@ -54,7 +58,8 @@ public class CustomerController {
     }
 
     @GetMapping("/customer")
-    public String getCustomer(@RequestParam(name = "id", required = true) String id, Model model) {
+    public String getCustomer(@RequestParam(name = "id", required = true) String id, Model model, HttpSession session) {
+        model.addAttribute(CsrfHttpSessionListener.TOKEN, session.getAttribute(CsrfHttpSessionListener.TOKEN).toString());
         model.addAttribute("customer", customerRepository.getCustomer(id));
         model.addAttribute("addresses", customerRepository.getAddresses(id));
         return "customer";
@@ -67,28 +72,52 @@ public class CustomerController {
     }
 
     @PostMapping("/api/customer/update-customer")
-    public String updateCustomer(CustomerUpdate customerUpdate, Model model) {
+    public String updateCustomer(CustomerUpdate customerUpdate, Model model, HttpSession session, @RequestParam("CSRFToken") String CSRFToken) throws AccessDeniedException {
+        String sessionToken = session.getAttribute(CsrfHttpSessionListener.TOKEN).toString();
+
+        if (!sessionToken.equals(CSRFToken)) {
+            throw new AccessDeniedException("Forbidden Access");
+        }
+
         customerRepository.updateCustomer(customerUpdate);
         customersAndRestaurants(model);
         return "/customers-and-restaurants";
     }
 
     @DeleteMapping("/customer/address")
-    public String deleteCustomerAddress(@RequestParam(name = "id", required = true) String id) {
+    public String deleteCustomerAddress(@RequestParam(name = "id", required = true) String id, HttpSession session, @RequestParam("CSRFToken") String CSRFToken) throws AccessDeniedException {
+        String sessionToken = session.getAttribute(CsrfHttpSessionListener.TOKEN).toString();
+
+        if (!sessionToken.equals(CSRFToken)) {
+            throw new AccessDeniedException("Forbidden Access");
+        }
+
         int identificator = Integer.valueOf(id);
         customerRepository.deleteCustomerAddress(identificator);
         return "/customers-and-restaurants";
     }
 
     @PostMapping("/api/customer/address/update-address")
-    public String updateCustomerAddress(Address address, Model model) {
+    public String updateCustomerAddress(Address address, Model model, HttpSession session, @RequestParam("CSRFToken") String CSRFToken) throws AccessDeniedException {
+        String sessionToken = session.getAttribute(CsrfHttpSessionListener.TOKEN).toString();
+
+        if (!sessionToken.equals(CSRFToken)) {
+            throw new AccessDeniedException("Forbidden Access");
+        }
+
         customerRepository.updateCustomerAddress(address);
         customersAndRestaurants(model);
         return "/customers-and-restaurants";
     }
 
     @PostMapping("/customer/address")
-    public String putCustomerAddress(NewAddress newAddress, Model model){
+    public String putCustomerAddress(NewAddress newAddress, Model model, HttpSession session, @RequestParam("CSRFToken") String CSRFToken) throws AccessDeniedException {
+        String sessionToken = session.getAttribute(CsrfHttpSessionListener.TOKEN).toString();
+
+        if (!sessionToken.equals(CSRFToken)) {
+            throw new AccessDeniedException("Forbidden Access");
+        }
+
         customerRepository.putCustomerAddress(newAddress);
         customersAndRestaurants(model);
         return "/customers-and-restaurants";
